+++ CHANGELOG.md
@@ -46,6 +46,7 @@ Changelog
   This is a known bug in the Kotlin compiler, follow https://github.com/ZacSweers/metro/issues/1556
   ```
 - [IR] Avoid generating unnecessary `Provider` refcounts for bindings only used by graph injector functions.
+- [IR] When reporting graph failures in dynamic graphs, report the original call location in error reporting.
 
 ### Fixes
 
@@ -56,6 +57,7 @@ Changelog
 - [IR] Check for `open` or `final` modality on classes first before searching for injectable constructors.
 - [IR] Mark all supertypes' member injections as dependencies of subtypes' member injections. That's a wordier way of saying that member-injecting a class `Dog` that extends `Animal` also depends on member-injected dependencies of `Animal`.
 - [IR/Native] Work around `Abstract function '___' is not implemented in non-abstract class 'Impl'` info warnings from native compilations.
+- [IR/Dynamic Graphs] Always check if a key is replaced by a dynamic type key when building binding graphs. Previously, some bindings would be added even if a dynamic key was available, resulting in unexpected `DuplicateBinding` errors.
 - [IC] Record lookups of contributed classes when looking up hints from IR. Previously Metro only recorded a lookup of the generated hint function, which appears to not be enough for Kotlin 2.3.20.
 - [IC] Link IR-generated hint function files back to source class via expect-actual tracker to link their compilations. This fixes an edge case where simply changing a contribution scope (or removing it) could leave behind a stale hint file that downstream compilations would incidentally read.